ディレクトリの配置を整理しました。 - 当初のサンプルアプリは/sampleに移動 - 私が開発した最新版のアプリはmain_app.pyという名前でルートに移動
