--- personal/README.md
@@ -1,1 +1,0 @@
-# 各人の作業用ディレクトリ
